+++ spec/app/models/wiki_page.rb
@@ -5,7 +5,10 @@ class WikiPage
 
   field :title, :type => String
   field :transient_property, :type => String, :versioned => false
+  field :author, :type => String
   max_versions 5
+
+  attr_protected :author
 
   has_many :comments, :dependent => :destroy, :validate => false
   has_many :child_pages, :class_name => "WikiPage", :dependent => :delete